+++ dbms/src/Common/Spiller.cpp
@@ -15,6 +15,7 @@
 #include <Common/Spiller.h>
 #include <DataStreams/NativeBlockOutputStream.h>
 #include <DataStreams/NullBlockInputStream.h>
+#include <DataStreams/SpilledFilesInputStream.h>
 #include <IO/CompressedWriteBuffer.h>
 #include <IO/WriteBufferFromFile.h>
 #include <Poco/Path.h>
@@ -36,41 +37,6 @@ SpilledFile::~SpilledFile()
     catch (...)
     {
     }
-}
-
-SpilledFilesInputStream::SpilledFilesInputStream(const std::vector<String> & spilled_files_, const Block & header_)
-    : spilled_files(spilled_files_)
-    , header(header_)
-{
-    RUNTIME_CHECK_MSG(!spilled_files.empty(), "Spilled files must not be empty");
-    current_reading_file_index = 0;
-    current_file_stream = std::make_unique<SpilledFileStream>(spilled_files[0], header);
-}
-
-Block SpilledFilesInputStream::readImpl()
-{
-    Block ret;
-    ret = current_file_stream->block_in->read();
-    if (ret)
-        return ret;
-    current_reading_file_index++;
-    for (; current_reading_file_index < spilled_files.size(); current_reading_file_index++)
-    {
-        current_file_stream = std::make_unique<SpilledFileStream>(spilled_files[current_reading_file_index], header);
-        ret = current_file_stream->block_in->read();
-        if (ret)
-            return ret;
-    }
-    return ret;
-}
-
-Block SpilledFilesInputStream::getHeader() const
-{
-    return header;
-}
-String SpilledFilesInputStream::getName() const
-{
-    return "SpilledFiles";
 }
 
 Spiller::Spiller(const String & id_, bool is_input_sorted_, size_t partition_num_, const String & spill_dir_, const Block & input_schema_, LoggerPtr logger_)
@@ -133,10 +99,14 @@ BlockInputStreams Spiller::restoreBlocks(size_t partition_id, size_t max_stream_
     {
         for (const auto & file : spilled_files[partition_id])
         {
-            if likely (file->exists())
+            if (likely(file->exists()))
             {
                 std::vector<String> files{file->path()};
                 ret.push_back(std::make_shared<SpilledFilesInputStream>(files, input_schema));
+            }
+            else
+            {
+                LOG_WARNING(logger, "Spill file {} does not exists", file->path());
             }
         }
     }
@@ -144,14 +114,16 @@ BlockInputStreams Spiller::restoreBlocks(size_t partition_id, size_t max_stream_
     {
         size_t return_stream_num = std::min(max_stream_size, spilled_files[partition_id].size());
         std::vector<std::vector<String>> files(return_stream_num);
-        for (size_t i = 0; i < spilled_files[partition_id].size(); i++)
+        for (size_t i = 0; i < spilled_files[partition_id].size(); ++i)
         {
-            if likely (spilled_files[partition_id][i]->exists())
+            if (likely(spilled_files[partition_id][i]->exists()))
                 files[i % return_stream_num].push_back(spilled_files[partition_id][i]->path());
+            else
+                LOG_WARNING(logger, "Spill file {} does not exists", spilled_files[partition_id][i]->path());
         }
-        for (size_t i = 0; i < return_stream_num; i++)
+        for (size_t i = 0; i < return_stream_num; ++i)
         {
-            if likely (!files[i].empty())
+            if (likely(!files[i].empty()))
                 ret.push_back(std::make_shared<SpilledFilesInputStream>(files[i], input_schema));
         }
     }
